Save uploaded into custom temp directory
Previously, the uploaded files were saved into the system temporary
directory, however, this might cause "Invalid Cross-devices Link" error
if the temporary directory locating at another disk partition. Thus, we
now save the files in our custom temporary directory, which is under the
current directory.

Also, we won't persist the uploaded file now as it seems that this is
not necessary.