Finish the upload handler
The whole handler is now implemented. It first receive the uploaded
video, and then initializes the analyzer and invoke inference script at
a predefined location (we might make this configurable in the future).
After that, we read the file containing inference results at a
predefined location, and then processes and propagates them to the
frontend.

We also change the default port for our HTTP server from 8080 to 9090
since binding the former sometimes fails since it is already in use.