Refactor analyzer module and visibility settings
We combine `inference` module into `analyzer`, since the former is
supposed to only be used by the latter only. This also explains why
`inference` module is defined as a private module of `analyzer`.

We also refine the item visibilities to avoid excessive uses of
`pub(crate)`.

Finally, we add some document for the code in `analyzer` module.